integrate arduino version of DF class
